Extract sanitize file into seperate module
