+++ tests/test_sanitizer.py
@@ -4,4 +4,6 @@ from repobee_sanitizer import sanitizer
 
 def test_register():
     """Just test that there is no crash"""
-    repobee.try_register_plugin(sanitizer, sanitizer.SanitizeRepo)
+    repobee.try_register_plugin(
+        sanitizer, sanitizer.SanitizeRepo, sanitizer.SanitizeFile
+    )
